+++ Vic3ToHoI4Tests.vcxproj
@@ -80,7 +80,7 @@
       <MultiProcessorCompilation>true</MultiProcessorCompilation>
       <ProgramDataBaseFileName>$(SolutionDir)lib\common.pdb</ProgramDataBaseFileName>
       <ForcedIncludeFiles>pch.h</ForcedIncludeFiles>
-      <PrecompiledHeader>Create</PrecompiledHeader>
+      <PrecompiledHeader>NotUsing</PrecompiledHeader>
       <PrecompiledHeaderFile>pch.h</PrecompiledHeaderFile>
       <PrecompiledHeaderOutputFile>$(SolutionDir)lib\pch.pch</PrecompiledHeaderOutputFile>
       <DebugInformationFormat>EditAndContinue</DebugInformationFormat>
@@ -98,6 +98,7 @@
     <ClCompile Include="external\commonItems\external\googletest\googletest\src\gtest-all.cc" />
     <ClCompile Include="external\commonItems\external\googletest\googletest\src\gtest_main.cc" />
     <ClCompile Include="src\hoi4_world\countries\hoi4_country_converter_tests_tech.cpp" />
+    <ClCompile Include="src\hoi4_world\states\hoi4_states_converter_tests_2.cpp" />
     <ClCompile Include="src\mappers\world\world_mapper_builder_tests.cpp" />
     <ClCompile Include="src\vic3_world\countries\vic3_country_tests.cpp" />
     <ClCompile Include="src\vic3_world\ideologies\ideologies.cpp" />
